Ejercicio 4.21

+++ part4/controllers/blogs.js
@@ -46,6 +46,27 @@ blogsRouter.get('/', async (request, response) => {
 
 
   blogsRouter.delete('/:id', async (request, response) => {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+   
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+
+    if (!blog) {
+      return response.status(404).json({ error: 'Blog not found' })
+  }
+
+    if (!blog.user) {
+      return response.status(400).json({ error: 'User not found in blog' })
+  }
+
+    if (blog.user.toString() !== user.id.toString() ){
+      return response.status(403).json({ error: 'User incorrect' })
+    }
+  
+
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
